+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+2.0.2
+* added support for Google Advertising Id
+
 2.0.1
 * quieted logging by default
 * fixed intermittent crash on older devices
 * added ability to disable geo-targeting for apps that already use location services
+* fixed intermittent crash when displaying video ad
 * improved user agent reporting
 * deprecated AdPrompts in conformance of Google policies